fix upload-unbound-stream tests breaking on windows
